tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.transform.1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.transform.1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.transform.1</h1>
<p class="desc">Shadows take account of transformations</p>


<script>
var t = async_test("Shadows take account of transformations");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.shadowOffsetY = 50;
  ctx.shadowColor = '#0f0';
  ctx.translate(100, 100);
  ctx.fillRect(-100, -150, 100, 50);

  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
